Avoid double "texification"
Because we use `Question.comments` in the LaTeX template, we
essentially called `texify_comments` twice for the penalty
comments. This caused markdown code formatting (`...`)
to become \textbackslashtexttt.
This change does remove the call to remove_ta_internal_comments,
but since the penalty comments are not recorded on Submitty, I
assumed that this was fine.